Adding monotonically increasing index test
This test creates an entity with monotonically increasing indices. This
should put pressure on the index writes which should all occur in the
same part of the index overloading a single server.

It should be noted that our current tests do not simulate parallel load.
This may mean that this overload does not appear in this test.