Added load/:size/*url url match
This url is used to generate parallel load by generating 'size' many
http requests to 'url'.

Additionally we now have the

/del/:kind

which is used to delete and time a small back of 'kind' entities from
the datastore. This replaces the multitude of custom urls we had
previously.

/clear/:kind

which is used to delete every entity of 'kind' in the datastore. This is
a reset mechanism and is not timed.