Adding control URL
The control does no work, but it does time how long it takes to do no
work. This URL is used to get a baseline with which we can compare other
actions.

It also can be used to test how fast the load/ URL is able to perform
URL fetches.