Added URL fetch timing info
Also fixed url prefex typo.

The result of initial testing indicates that the load test is able to
fetch at least 500 URLs simultaneously. The total time taken is
typically slightly more than the time taken by the slowest request.

An interesting side note is that it is possible to increase the max
number if concurrent requests served by an app engine instance.
Increasing this above 10, the default, appears to cause the control/ URL
to slow down indicating that even when the server is doing no work it
can't serve 500 concurrent requests efficiently.